rename router, add null checker for paths and methods
